Rename oniguruma-jni packages and module
