+++ oniguruma-jni/src/main/kotlin/me/zolotov/oniguruma/jni/OnigurumaLoader.kt
@@ -1,4 +1,4 @@
-package me.zolotov.oniguruma
+package me.zolotov.oniguruma.jni
 
 import java.nio.file.Files
 import java.nio.file.Path
